Support passing labels through existing transforms
This is the first step in supporting labels in the transformation
pipeline. All transforms now inherit from a `Transform` class that
handle passing frames to the transform implementation and passing the
label through the transform untouched. This might later form the basis
for `target_transforms` although the internal API of `Tranform` needs to
be cleaned up so that `_gen_params` isn't so disgusting.

Also added missing tests for

- ResizeVideo
- TimeApply
- RandomResizedCropVideo