Integrate resnet3d models from pretorched-x
All the work for these models is by Alex Andonian
(https://github.com/alexandonian/pretorched-x). The only changes I've
made here are to clean up the inflation code somewhat, and make the
models compatible with pytorch 1.1.0 (see details in
https://github.com/Cadene/pretrained-models.pytorch/issues/120) by
avoiding the use of `types.MethodType` that was present in
`modify_resnet`. There is no need to have this dynamic modification as
we own the model, so I've simply added the `features`, `logits` and
`last_linear` layers to the base model `ResNet3D` itself.

Currently this code isn't tested, but this forms the starting point of
testing and cleaning up the pretorched-x code so we can neatly separate
it from being integrated within `pretrainedmodels`, and rather only use
that as a dependency.